Refactor ignore certs
