+++ implants/lib/transport/src/http.rs
@@ -13,58 +13,8 @@ use std::sync::{
 #[cfg(feature = "doh")]
 use crate::dns_resolver::doh::DohProvider;
 
-use rustls_0_21::client::{HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier};
+use crate::tls_utils::legacy::AcceptAllCertVerifier;
 use std::str::FromStr;
-use std::time::SystemTime;
-
-#[derive(Debug)]
-struct AcceptAllCertVerifier;
-
-impl ServerCertVerifier for AcceptAllCertVerifier {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &rustls_0_21::Certificate,
-        _intermediates: &[rustls_0_21::Certificate],
-        _server_name: &rustls_0_21::ServerName,
-        _scts: &mut dyn Iterator<Item = &[u8]>,
-        _ocsp_response: &[u8],
-        _now: SystemTime,
-    ) -> Result<ServerCertVerified, rustls_0_21::Error> {
-        Ok(ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &rustls_0_21::Certificate,
-        _dss: &rustls_0_21::DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, rustls_0_21::Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &rustls_0_21::Certificate,
-        _dss: &rustls_0_21::DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, rustls_0_21::Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<rustls_0_21::SignatureScheme> {
-        vec![
-            rustls_0_21::SignatureScheme::ECDSA_NISTP384_SHA384,
-            rustls_0_21::SignatureScheme::ECDSA_NISTP256_SHA256,
-            rustls_0_21::SignatureScheme::ED25519,
-            rustls_0_21::SignatureScheme::RSA_PSS_SHA512,
-            rustls_0_21::SignatureScheme::RSA_PSS_SHA384,
-            rustls_0_21::SignatureScheme::RSA_PSS_SHA256,
-            rustls_0_21::SignatureScheme::RSA_PKCS1_SHA512,
-            rustls_0_21::SignatureScheme::RSA_PKCS1_SHA384,
-            rustls_0_21::SignatureScheme::RSA_PKCS1_SHA256,
-        ]
-    }
-}
 
 /// gRPC frame header utilities for encoding/decoding wire protocol frames
 mod grpc_frame {
