+++ implants/lib/transport/src/lib.rs
@@ -3,6 +3,9 @@ use pb::c2::transport::Type as TransportType;
 use pb::{c2::*, config::Config};
 use std::sync::mpsc::{Receiver, Sender};
 
+#[cfg(any(feature = "grpc", feature = "http1"))]
+mod tls_utils;
+
 #[cfg(feature = "grpc")]
 mod grpc;
